-------------------------------------------------------------------------------- Changeset Id: 864959 Date: 5/8/2009 6:21:21 PM
(dinov) This change is removing the binder state class and updating all Python binders to hold onto a PythonContext instead of a BinderState which holds onto a CodeContext which holds onto a PythonContext.  
 
The original idea behind BinderState was there was 1 / module and it would enable us to not have to flow context everywhere.  But because sites can be shared across multiple scopes that didn’t work out so well.  Now it’s a hinderance because the atomization is happening on the binderstate so we’re not sharing sites across modules.  It’s also breaking Harry’s __clrtype__ experimentation.  So now it’s just gone and we always use the context which is flowed in anyway.
 
There’s lots of little changes throughout the code base but most if it’s pretty straight forward.  The common theme is that we get a PythonContext and don’t need to access it’s DefaultBinderState property or we get the SharedContext off of the PythonContext instead of getting it from the DefaultBinderState.
 
Modules lose the initialization of the binder state code and the helpers are gone from PythonOps. 
 
All of the old code in BinderState is pulled into PythonContext unchanged.
 
In order to avoid flowing CodeContext into GetIndex sites I have also changed how generic methods work.  Instead of making all methods indexable and hiding it sometime there is a new subclass of BuiltinFunction called GenericBuiltinFunction.  It adds the indexing onto the BuiltinFunction.  I’ve been wanting to do this for a while but this is finally a good execute to do it – we now get the right error message on [].append[int]   We should also do the same w/ PythonType at some point.  That required an update to the restrictions in MetaBuiltinFunction to account for the type of built-in function when producing the rule.
 
I’ve also extended the PythonGetMemberBinder fast path.  It now supports NamespaceTracker objects and supports Scope more fully.
 


(Shelveset: RemoveBinderStateFinal2;REDMOND\dinov | SNAP CheckinId: 8296)
--------------------------------------------------------------------------------
Changeset Id: 864850
Date: 5/8/2009 5:14:30 PM

(dfugate) When support was removed for COM typelibs, the COM interop tests were not updated accordingly.  This shelveset:
* removes all typelib-only paths throughout the tests
* updates a number of bug IDs


(Shelveset: COM_TESTS;REDMOND\dfugate | SNAP CheckinId: 8295)