-------------------------------------------------------------------------------- Changeset Id: 1135813 Date: 9/17/2009 4:01:59 PM
(sborde) Updates to the DynamicWebServiceHelpers sample
Updated the sources in Merlin\Main\Languages\IronPython\Samples\DynamicWebServiceHelpers\sources to be exactly the same as in Merlin\Main\Libraries\DynamicWebServiceHelpers. Ideally, we will not have two copies, but I am not dealing with that right now. Also, removed reference to IronPython.dll make the library language-independent again.
Deleted the *.py sampels in Merlin\Main\Libraries\DynamicWebServiceHelpers so that we have only one copy atleast of these samples. Also, changed _test.py to test the *.py samples
Deleted the amazon.py sample as the web request now needs to be signed with a secret key, and its not worth dealing with this.
Added a Bing sample
Renamed math.py to mathservice.py so that it does not clash with the builtin math module. Also, changed the webservice url as the previous one is no longer valid.
Deleted the stale JSX tests for DynamicWebServiceHelpers



(Shelveset: DWSH;REDMOND\sborde | SNAP CheckinId: 9443)
--------------------------------------------------------------------------------
Changeset Id: 1115991
Date: 9/11/2009 11:01:30 AM

(dinov) Removes IronPython’s usage of SymbolId’s except for implementing IAttributesCollection on PythonDictionaries.  This is so that these can be obsolete when they’re moved to the new Microsoft.Dynamic.dll (where they need to remain for a release for reasonable backwards compatibility w/ existing hosts).

Pretty straight forward – just goes through all the code and removes the various transformations to and from SymbolId’s all over the place.




(Shelveset: RemoveSymbolIdsFinal;REDMOND\dinov | SNAP CheckinId: 9391)