-------------------------------------------------------------------------------- Changeset Id: 1227952 Date: 10/21/2009 9:57:12 PM
(sborde) CLR exception should include Python exception class name even for user exception types. We were leaving it empty because BaseException._message is initialized to "", whereas we were checking for null.
Also fixes a bug in sortvsmdi.py which is exposed only if run with IPy 2.X. Snap uses IPy 1.0 which does not happen to have an issue with the incorrect code. We made a breaking change in IPy 2.X which exposes the issue.



(Shelveset: msg;REDMOND\sborde | SNAP CheckinId: 9652)