-------------------------------------------------------------------------------- Changeset Id: 1346991 Date: 12/15/2009 11:56:36 AM
Fixes _ctypes union support so that it now works.  Also fixes _buffer_info API so that it can work (and enables one test which was previously failing).  _buffer_info is a little bit odd in that sometimes we need to save the buffer format when a type is created and other times we don’t (the tests cover this).  Union is easy – it just needs the same logic for getting/setting values that structures have.

GeneratorRewriter is looking for something which isn’t a DelayedTupleExpression.  But we can get back a simple FieldExpression for the LHS and when that happens the cast to BlockExpression fails.  So instead we should always test that we have a BlockExpression.




(Shelveset: CtypesUnionAndGeneratorFullFramesBugFinal;REDMOND\dinov | SNAP CheckinId: 10020)
--------------------------------------------------------------------------------
Changeset Id: 1346865
Date: 12/15/2009 10:24:40 AM

Removes some excess methods from BigIntegerV4 that the underlying BigInt does not implement, to help phase out this wrapper:
  - Moves ToFloat64 and TryToFloat64 into MathUtils as extension methods
  - Removes the IConvertible interface from BigIntegerV4

Fixes minor bugs in math.cs and socket.cs, and removes dependenies on BigInteger's IConvertible methods.

Corrects BigIntegerV2's broken right-shift so we don't have to permute the value in LongOps.RightShift (which is broken given BigIntegerV4's correct behavior).

Fixes BigIntegerV4's float and double conversions to overflow properly by adding an implicit BigInteger to double conversion.



(Shelveset: BigIntV4;REDMOND\ddicato | SNAP CheckinId: 10019)