-------------------------------------------------------------------------------- Changeset Id: 1481710 Date: 1/22/2010 2:30:19 PM
Adds /notification option to Chiron.  This option causes Chiron to pop a notification ballon on startup identifying where it is running.  It then provides a content menu for stopping the server, bringing up the status, or opening a web browser which points at the server.  



(Shelveset: ChironNotification;REDMOND\dinov | SNAP CheckinId: 10179)