-------------------------------------------------------------------------------- Changeset Id: 1494245 Date: 1/26/2010 3:33:49 PM
(ddicato) CLR4 only: Track in PythonContext when Assembly.LoadFile is called. Change AssemblyResolve hook to skip searching sys.path when the RequestingAssembly is not loaded in the current context.

(dfugate) Added an env. var to iptest.test_env to see if we're using the DLR that comes installed with .NET 4.0.



(Shelveset: cp23506;REDMOND\ddicato | SNAP CheckinId: 10206)