-------------------------------------------------------------------------------- Changeset Id: 1525040 Date: 1/28/2010 3:49:13 PM
Fixes a regression in 2.6.1 found by the full test run.  This issue is showing up now because of the lazy production of the ASTs.  But it’s been there for a while. But here’s what’s happening:

First off FunctionCode.GetAlLCode is broken – it’s losing the list of function code objects by exchanging in curCodeList (which is null) instead of initialCode.  This causes us not to flip tracing onto a code object.  After fixing that we flip the function codes of more objects which exposes an issue of UpdateDelegate compiling code which causes the creation of more FunctionCode objects.  That causes us to re-enter and corrupt our lock which is held while we’re updating the function codes. So now we won’t compile function code objects switching back from tracing on to off just like we don’t when switching from off to on.

GetRootCodeNoUpdating was also broken (and would spin forever if we hit this race) although that’s unrelated to the failures we’re seeing 




(Shelveset: Fix261Regression;REDMOND\dinov | SNAP CheckinId: 10227)