-------------------------------------------------------------------------------- Changeset Id: 1551907 Date: 2/3/2010 12:47:59 AM
Adds a copy of .NET 4’s IStructuralEquatable and IStructuralComparable interfaces to the CLR2 version, and implements these interfaces on many of our Python types. Changes resolution of __hash__, __eq__, __ne__, __gt__, __lt__, __ge__, __le__ to look for these interfaces first, as well as IComparable later (after OperatorResolver) if the type is non-primitive.

In CLR4, we eliminate the IValueEquality interface entirely, replacing it with IStructuralEquatable where appropriate. It is kept in CLR2 only to avoid breaking changes.

Also fixes some broken equality and comparison behavior in dict, list, and tuple when __len__ is overridden.




(Shelveset: cp23338;REDMOND\ddicato | SNAP CheckinId: 10269)