-------------------------------------------------------------------------------- Changeset Id: 1591043 Date: 2/13/2010 7:30:01 AM
- Bug 359739 (test_name.py) is fixed
- CP23878:  added a regression.  Fixed
- CP23914:  added a regression.  Fixed
- CP23992:  added a regression.  Not fixed
- CP24169:  added a regression.  Fixed
- CP24077:  added a regression.  Fixed


(Shelveset: CP03;REDMOND\dfugate | SNAP CheckinId: 10360)
--------------------------------------------------------------------------------
Changeset Id: 1590796
Date: 2/12/2010 6:36:38 PM

Modifies calls to “unicode(…)” so that they behave like calls to unicode on CPython 2.6.  When we see a call to a NameExpression where the name is “unicode” we generate an in-line check to see if the object we’re calling is actually str.  If it is then we generate a call to a set of methods defined on a UnicodeHelper class.  These methods do the same thing as CPython’s unicode function:
	Call __unicode__ first if it’s defined, otherwise try __str__ when called with just oen param
	Decode strings/bytes/buffers/etc if an encoding type is provided when called with >1 param.

Also includes  a change to make ipyw.exe on Dev10 be a real windows app.







(Shelveset: UnicodeStr;REDMOND\dinov | SNAP CheckinId: 10353)
--------------------------------------------------------------------------------
Changeset Id: 1590569
Date: 2/12/2010 3:25:41 PM


Fixes a bug where addressof() returns the address of a function instead the address where the address of the function is stored.



(Shelveset: FixCtypesAddressBug;REDMOND\dinov | SNAP CheckinId: 10345)
--------------------------------------------------------------------------------
Changeset Id: 1590268
Date: 2/12/2010 11:53:12 AM

- CodePlex 20370: added regression.  Fixed
- CodePlex 23562: added regression.  Fixed
- CodePlex 18595: added regression.  Fixed



(Shelveset: CP03;REDMOND\dfugate | SNAP CheckinId: 10343)
--------------------------------------------------------------------------------
Changeset Id: 1591043
Date: 2/13/2010 7:30:01 AM

- Bug 359739 (test_name.py) is fixed
- CP23878:  added a regression.  Fixed
- CP23914:  added a regression.  Fixed
- CP23992:  added a regression.  Not fixed
- CP24169:  added a regression.  Fixed
- CP24077:  added a regression.  Fixed


(Shelveset: CP03;REDMOND\dfugate | SNAP CheckinId: 10360)