-------------------------------------------------------------------------------- Changeset Id: 1605388 Date: 2/24/2010 12:42:49 PM
Fixes a parser bug related to multi-line strings w/ state.  Currently when we hit a line like:
	“””foo\

We will report an error because of the EOF (we’re parsing one line at a time here) in the middle of the string.  Now if we’re in verbatim mode we’ll return an IncompleteStringToken and set our state so we can resume parsing on the next line.

This was exposed via IronPython Tools.




(Shelveset: FixMultiLineParserBug;REDMOND\dinov | SNAP CheckinId: 10394)