-------------------------------------------------------------------------------- Changeset Id: 1613348 Date: 3/1/2010 11:49:15 PM
Switches underlying implementation of bigint and complex to System.Numerics

Also fixes a couple CodePlex bugs:
 - 25893: <float> in <bytearray> throws wrong exception
 - 25895: coerce(<complex>, <huge_long>) doesn't throw




(Shelveset: numerics;REDMOND\ddicato | SNAP CheckinId: 10430)
--------------------------------------------------------------------------------
Changeset Id: 1612839
Date: 3/1/2010 4:00:48 PM

Updates the version numbers for 2.6.1 in the main branch. This is a superset of the version number update that went into the servicing branch for the 2.6.1 RC, with a couple exceptions:
 - IronPython.wxs: updated product version with today’s date instead of 02/02 so that this version is distinct from the RC.
 - DLR-related AssemblyInfo.cs: updated AssemblyVersion to 1.0.0.0. This was not necessary on my last update to the servicing branch because it had already been changed to 1.0.0.0

In general, any DLR AssemblyVersion changed from 0.9.6.20 to 1.0.0.0 and Assembly[File/Informational]Version from 1.0 to 1.0.1. Also updated any IronPython DisplayName to “IronPython 2.6.1” and Version to 2.6.10920.0 and re-generated ETScenarios\ETScenariosCSLinq[V4].dll.




(Shelveset: 261VersionNumbers;REDMOND\ddicato | SNAP CheckinId: 10426)
--------------------------------------------------------------------------------
Changeset Id: 1612839
Date: 3/1/2010 4:00:48 PM

Updates the version numbers for 2.6.1 in the main branch. This is a superset of the version number update that went into the servicing branch for the 2.6.1 RC, with a couple exceptions:
 - IronPython.wxs: updated product version with today’s date instead of 02/02 so that this version is distinct from the RC.
 - DLR-related AssemblyInfo.cs: updated AssemblyVersion to 1.0.0.0. This was not necessary on my last update to the servicing branch because it had already been changed to 1.0.0.0

In general, any DLR AssemblyVersion changed from 0.9.6.20 to 1.0.0.0 and Assembly[File/Informational]Version from 1.0 to 1.0.1. Also updated any IronPython DisplayName to “IronPython 2.6.1” and Version to 2.6.10920.0 and re-generated ETScenarios\ETScenariosCSLinq[V4].dll.




(Shelveset: 261VersionNumbers;REDMOND\ddicato | SNAP CheckinId: 10426)