-------------------------------------------------------------------------------- Changeset Id: 1760502 Date: 5/3/2010 2:40:28 PM
26499 – Need to track whether a type inherits from ABCMeta and check all abstracts are overridden
Added flags to PythonType, one that is set when __abstractmethods__ is defined, and another when it is set to be a non-empty iterator of strings. Changed instantiation behavior to check these flags and raise an exception if any abstract methods exist on the type.

20279 – Optimization: (2,3,4)[0:] should return same instance
Fixed, test added.

23663 – Trivial: Perf: potential perf optimizations
Most of these refer to BigInteger, which is being phased out in favor of .NET 4.0, or files/comments that otherwise no longer exist. Two issues in this bug still seem to be an issue today – the locking behavior of PythonFile and the magic number 20 for list initialization. Changed list initialization value from magic number to const int, although our choice of 20 is still arbitrary and might be better off as a lower number. Changed PythonFile operations that take/release locks more than once to take/release a single lock when possible.

23725 – Trivial: "typeerror "was not raised as expected in test_coercion.py when eval(3.0*(2,)) in 32bit vista
Already fixed, enabled test

23701 – Trivial: remove sys.getrefcount
Already fixed, enabled remaining test (test_socket)

23746 – Trivial: IndexError: Index was outside the bounds of the array
Fixed original error. Remaining failure is dupe of 23750

24754 – test_handlers in test_codecs is failing
Original error was already fixed. Remaining failure is dupe of 23750

25419 – test_container_iterator in test_dict.py disabled
Made required test changes in test_dict and test_deque, and re-enabled tests




(Shelveset: 262Bugs01;REDMOND\ddicato | SNAP CheckinId: 10712)