-------------------------------------------------------------------------------- Changeset Id: 1773679 Date: 5/11/2010 2:44:26 AM
Adds CPython 2.7's standard lib to the repository - one original copy and one that is merged with our changes to the 2.6 tests.

Also adds auto-numbering support of field names in format strings to enable test_support.py to work.




(Shelveset: 27Lib;REDMOND\ddicato | SNAP CheckinId: 10771)
--------------------------------------------------------------------------------
Changeset Id: 1773536
Date: 5/11/2010 1:04:10 AM

Re-order some checks in our dictionary implementation so when a key matches we’ll notice is sooner and avoid some if-checks.

Add a missing optimization to the importer – we should publish a null indirection entry when a nested import fails so that we don’t attempt to import it over and over again.  This results in a 5x improvement on the Django Unladen Swallow benchmark.  Also some small optimizations to avoid creating unnecessary garbage while importing.



(Shelveset: dictandimporttweaks2;REDMOND\dinov | SNAP CheckinId: 10769)