-------------------------------------------------------------------------------- Changeset Id: 1829946 Date: 5/24/2010 4:29:18 PM
Reduces memory usage of AST nodes.  This replaces our current storing of index, line, column with just storing an index.  To resolve the line and column we store the line start information in the PythonAst which we can access via the parent chain.

Lots of places are touched but it’s a pretty straight forward transformation.  The interesting changes are:
                We no longer use TokenizerBuffer, it’s purpose was to track the line information
                We now need to construct the PythonAst before constructing the child nodes.  This is because we need to assign some parent node as we set the indexes otherwise you can’t access line information from unbound trees.
                We need to accommodate updating the line mapping information.  We currently just do a pass at the end of parsing to transform all of the line info.  For error reporting we use Tokenizer’s IndexToLocation which also goes to the source unit for line updates.


Adds IndexSpan to the DLR and also adds HasLineMapping to source unit.




(Shelveset: AstMemoryFinal5;REDMOND\dinov | SNAP CheckinId: 10826)